add assets
