+++ Vendor-Invoice-Auditing-System-frontend/src/components/Error/NotFound.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Typography } from '@mui/material';
+
+const NotFound = () => {
+  return (
+    <Typography variant="h4">Page not found</Typography>
+  );
+}
+
+export default NotFound;
