--- Vendor-Invoice-Auditing-System-frontend/src/components/ErrorComponent/NotFound.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Typography } from '@mui/material';
-
-const NotFound = () => {
-  return (
-    <Typography variant="h4">Page not found</Typography>
-  );
-}
-
-export default NotFound;
