+++ Vendor-Invoice-Auditing-System-frontend/src/components/FooterComponent/Footer.jsx
@@ -10,15 +10,21 @@ const Footer = () => {
             <div className="footer-container">
                 <div className="footer-link">
                     <EmailIcon className="icon" />
-                    <a href="mailto:nisrine@nisrinekane.com">Email</a>
+                    <a href="mailto:nisrine@nisrinekane.com" target="_blank">
+                        Email
+                    </a>
                 </div>
                 <div className="footer-link">
                     <GitHubIcon className="icon" />
-                    <a href="https://github.com/nisrinekane">GitHub</a>
+                    <a href="https://github.com/nisrinekane" target="_blank">
+                        GitHub
+                    </a>
                 </div>
                 <div className="footer-link">
                     <LinkedInIcon className="icon" />
-                    <a href="https://www.linkedin.com/in/nisrinekane/">LinkedIn</a>
+                    <a href="https://www.linkedin.com/in/nisrinekane/" target="_blank">
+                        LinkedIn
+                    </a>
                 </div>
             </div>
             <p className="footer-copyright">&copy; {new Date().getFullYear()}</p>
