--- Vendor-Invoice-Auditing-System-frontend/src/components/IsometricLayout/IsometricLayout.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import './IsometricLayout.css';
-import FileUploadForm from '../FileUploadFormComponent/FileUploadForm';
-
-const IsometricLayout = () => {
-    return (
-        <div className="isometric-content">
-            <FileUploadForm />
-        </div>
-    );
-};
-
-export default IsometricLayout;
