+++ Vendor-Invoice-Auditing-System-frontend/src/components/Navbar/Navbar.jsx
@@ -1,10 +1,14 @@
 import React from 'react'
 import './Navbar.css'
+import logo from './logo.png'
 
 const Navbar = () => {
     return (
         <nav className="navbar">
-            <a href="/" className="navbar-logo">AutoAudit</a>
+            <div className="navbar-logo-title">
+                <img className="navbar-logo" src={logo} alt="logo" />
+                <a href="/" className="navbar-title">AutoAudit</a>
+            </div>
             <div className="navbar-links">
                 <a href="#features">Features</a>
                 <a href="#pricing">Pricing</a>
